+++ src/main/java/walkingkooka/spreadsheet/dominokit/find/SpreadsheetFindDialogComponentSpreadsheetParserTokenVisitor.java
@@ -19,17 +19,17 @@ package walkingkooka.spreadsheet.dominokit.find;
 
 import walkingkooka.spreadsheet.dominokit.textmatch.TextMatchComponent;
 import walkingkooka.spreadsheet.expression.function.SpreadsheetExpressionFunctions;
-import walkingkooka.spreadsheet.parser.ConditionSpreadsheetParserToken;
-import walkingkooka.spreadsheet.parser.EqualsSpreadsheetParserToken;
-import walkingkooka.spreadsheet.parser.GreaterThanEqualsSpreadsheetParserToken;
-import walkingkooka.spreadsheet.parser.GreaterThanSpreadsheetParserToken;
-import walkingkooka.spreadsheet.parser.LessThanEqualsSpreadsheetParserToken;
-import walkingkooka.spreadsheet.parser.LessThanSpreadsheetParserToken;
-import walkingkooka.spreadsheet.parser.NamedFunctionSpreadsheetParserToken;
-import walkingkooka.spreadsheet.parser.NotEqualsSpreadsheetParserToken;
-import walkingkooka.spreadsheet.parser.SpreadsheetParserToken;
-import walkingkooka.spreadsheet.parser.SpreadsheetParserTokenVisitor;
-import walkingkooka.spreadsheet.parser.TextSpreadsheetParserToken;
+import walkingkooka.spreadsheet.formula.ConditionSpreadsheetParserToken;
+import walkingkooka.spreadsheet.formula.EqualsSpreadsheetParserToken;
+import walkingkooka.spreadsheet.formula.GreaterThanEqualsSpreadsheetParserToken;
+import walkingkooka.spreadsheet.formula.GreaterThanSpreadsheetParserToken;
+import walkingkooka.spreadsheet.formula.LessThanEqualsSpreadsheetParserToken;
+import walkingkooka.spreadsheet.formula.LessThanSpreadsheetParserToken;
+import walkingkooka.spreadsheet.formula.NamedFunctionSpreadsheetParserToken;
+import walkingkooka.spreadsheet.formula.NotEqualsSpreadsheetParserToken;
+import walkingkooka.spreadsheet.formula.SpreadsheetParserToken;
+import walkingkooka.spreadsheet.formula.SpreadsheetParserTokenVisitor;
+import walkingkooka.spreadsheet.formula.TextSpreadsheetParserToken;
 import walkingkooka.text.cursor.parser.ParserToken;
 import walkingkooka.tree.expression.ExpressionFunctionName;
 import walkingkooka.visit.Visiting;
