+++ src/test/java/walkingkooka/spreadsheet/dominokit/condition/ConditionRightSpreadsheetParserTokenComponentTest.java
@@ -21,8 +21,8 @@ import elemental2.dom.HTMLFieldSetElement;
 import org.junit.jupiter.api.Test;
 import walkingkooka.reflect.JavaVisibility;
 import walkingkooka.spreadsheet.dominokit.value.FormValueComponentTesting;
+import walkingkooka.spreadsheet.formula.ConditionRightSpreadsheetParserToken;
 import walkingkooka.spreadsheet.meta.SpreadsheetMetadataTesting;
-import walkingkooka.spreadsheet.parser.ConditionRightSpreadsheetParserToken;
 
 import java.util.Optional;
