+++ src/test/java/walkingkooka/spreadsheet/dominokit/formula/SpreadsheetFormulaComponentExpressionParserFunctionTest.java
@@ -20,8 +20,8 @@ package walkingkooka.spreadsheet.dominokit.formula;
 import org.junit.jupiter.api.Test;
 import walkingkooka.spreadsheet.SpreadsheetErrorKind;
 import walkingkooka.spreadsheet.SpreadsheetFormula;
+import walkingkooka.spreadsheet.formula.SpreadsheetParsers;
 import walkingkooka.spreadsheet.meta.SpreadsheetMetadataTesting;
-import walkingkooka.spreadsheet.parser.SpreadsheetParsers;
 import walkingkooka.text.cursor.TextCursors;
 import walkingkooka.text.printer.TreePrintableTesting;
 import walkingkooka.util.FunctionTesting;
